requirements.txt
dnspython
boto3
slack-sdk
python-dotenv
